Modified search script and relocated database scriptinside script

+++ crawl/views.py
@@ -90,33 +90,50 @@ def crawl(request):
                 fl=os.path.basename(a.path) #parsing the filename from the url
                 ### NLP SPACY
                 String = "La Camera di Commercio territorialmente competente effettuerà controlli su un campione pari" #the search string from a pdf 
+                print('The search string is :\n',String)
                 pdf_reader= pdf.PdfFileReader(fl)
                 if pdf_reader.getIsEncrypted()=='False' and pdf_reader.getNumPages()>0:
+                    print('The condition is met for the pdf named:\n',fl)
                     NumPages = pdf_reader.getNumPages()
                     for i in range(0,NumPages):
                         PageObj = pdf_reader.getPage(i)
-                        print("this is page " + str(i))
+                        print("This is page " + str(i))
                         Text = PageObj.extractText() #Extracting the text from the pdf page and then using the nlp to find the matching pattersns in the later stage 
                         print(f'The extracted text from the page {i} is :\n{Text}')
                         ResSearch = re.search(String, Text)
-                    pass #This will only execute the unencrypted files and valid files
+                        if ResSearch =='True':
+                            file_size=response.headers.get('content-length', 0)
+                            content_type=response.headers.get('Content-Type', 0)
+                            last_modified=response.headers.get('Last-Modified', 0)
+                            expiry_date=response.headers.get('Expires', 0)
+                            cache_control=response.headers.get('Cache', 0)
+                            server=response.headers.get('Server', 0)
+
+                            pub_instance = Publisher.objects.create(name=fl,links=u,file_size=file_size,content_type=content_type,last_modified_y=last_modified,expiry_date_y=expiry_date,cache_control_y=cache_control,server_y=server) #creating the entry in the database
+                            pub_instance.save() #saving the info for each url to database
+
+                            for row in Publisher.objects.all().reverse(): #removing all the duplicate items from the database
+                                if Publisher.objects.filter(name=row.name).count() > 1: #using name as a filter
+                                    print('Found Duplicate item:\n',row.name)
+                                    row.delete()
+                    pass #This will only execute the unencrypted files and valid files which contain the search string
                 else:
                     break
                 ###
-                file_size=response.headers.get('content-length', 0)
-                content_type=response.headers.get('Content-Type', 0)
-                last_modified=response.headers.get('Last-Modified', 0)
-                expiry_date=response.headers.get('Expires', 0)
-                cache_control=response.headers.get('Cache', 0)
-                server=response.headers.get('Server', 0)
+                # file_size=response.headers.get('content-length', 0)
+                # content_type=response.headers.get('Content-Type', 0)
+                # last_modified=response.headers.get('Last-Modified', 0)
+                # expiry_date=response.headers.get('Expires', 0)
+                # cache_control=response.headers.get('Cache', 0)
+                # server=response.headers.get('Server', 0)
 
-                pub_instance = Publisher.objects.create(name=fl,links=u,file_size=file_size,content_type=content_type,last_modified_y=last_modified,expiry_date_y=expiry_date,cache_control_y=cache_control,server_y=server) #creating the entry in the database
-                pub_instance.save() #saving the info for each url to database
+                # pub_instance = Publisher.objects.create(name=fl,links=u,file_size=file_size,content_type=content_type,last_modified_y=last_modified,expiry_date_y=expiry_date,cache_control_y=cache_control,server_y=server) #creating the entry in the database
+                # pub_instance.save() #saving the info for each url to database
 
-                for row in Publisher.objects.all().reverse(): #removing all the duplicate items from the database
-                    if Publisher.objects.filter(name=row.name).count() > 1: #using name as a filter
-                        print('Found Duplicate item:\n',row.name)
-                        row.delete()
+                # for row in Publisher.objects.all().reverse(): #removing all the duplicate items from the database
+                #     if Publisher.objects.filter(name=row.name).count() > 1: #using name as a filter
+                #         print('Found Duplicate item:\n',row.name)
+                #         row.delete()
     except Exception as exc:
         pass
         return render(request, 'result.html', {'list':all_links}) #redirecting to the results template page
